+++ packages/contracts/src/hooks/index.ts
@@ -1,5 +1,6 @@
 // Copyright (c) 2026 ObjectStack contributors. Apache-2.0 license.
 
 import { contractHook } from '../objects/contracts_contract.hook';
+import { obligationHook } from '../objects/contracts_obligation.hook';
 
-export const allHooks = [contractHook];
+export const allHooks = [contractHook, obligationHook];