+++ packages/project/src/objects/pm_resource.object.ts
@@ -24,7 +24,7 @@ export const Resource = ObjectSchema.create({
       required: true,
       group: 'core',
     }),
-    person: Field.lookup('user', {
+    person: Field.lookup('sys_user', {
       label: 'Person',
       required: false,
       group: 'core',